backend/requirements.txt
Django==1.8.3
django-autoslug==1.8.0
djangorestframework==3.1.3
factory-boy==2.5.2
psycopg2==2.6.1
